fleshed out more Simple KML parsing  * ColorStyle: color  * Document: shared styles  * Feature: container & document  * IconStyle: icon & beginnings of automatically applying scale, heading, and color  * LineStyle: width  * Style: icon & line substyles fixed bug in object ID parsing
git-svn-id: file:///Users/incanus/Documents/Projects/Development%20Seed/Mapping/Repository/MapBoxiPadDemo/trunk@37 1a2663a8-3dcc-4292-817d-c9c41621312c

+++ Classes/SimpleKMLLineStyle.h
@@ -12,9 +12,9 @@
 
 @interface SimpleKMLLineStyle : SimpleKMLColorStyle
 {
-    NSUInteger width;
+    CGFloat width;
 }
 
-@property (nonatomic, assign) NSUInteger width;
+@property (nonatomic, assign) CGFloat width;
 
 @end
